enforce terminal size better when watching
set the scroll region to the size of the terminal you're watching when
you first start watching a stream. this typically happens automatically
when watching someone who opens and closes something like vim (since
resetting the scroll region means explicitly setting the scroll region
to the size of the whole terminal, which will be the size of the
watching terminal in this case), but this helps for when a stream first
starts broadcasting. also be sure to reset it when you stop watching.